refactor: move setup and assertions in 'testsupport'
