+++ test/e2e/usersignup_test.go
@@ -6,15 +6,12 @@ import (
 	"time"
 
 	"github.com/codeready-toolchain/api/pkg/apis/toolchain/v1alpha1"
-	"github.com/codeready-toolchain/toolchain-common/pkg/cluster"
 	"github.com/codeready-toolchain/toolchain-e2e/testsupport"
-	"github.com/codeready-toolchain/toolchain-e2e/testsupport/md5"
 	"github.com/codeready-toolchain/toolchain-e2e/wait"
 
 	uuid "github.com/satori/go.uuid"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 type userSignupIntegrationTest struct {
@@ -73,7 +70,7 @@ func (s *userSignupIntegrationTest) TestUserSignupApproval() {
 func (s *userSignupIntegrationTest) TestTargetClusterSelectedAutomatically() {
 	// Create user signup
 	s.setApprovalPolicyConfig("automatic")
-	userSignup := newUserSignup(s.T(), s.awaitility.Host(), "reginald@alpha.com", "reginald@alpha.com")
+	userSignup := testsupport.NewUserSignup(s.T(), s.awaitility.Host(), "reginald@alpha.com", "reginald@alpha.com")
 
 	// Remove the specified target cluster
 	userSignup.Spec.TargetCluster = ""
@@ -82,28 +79,28 @@ func (s *userSignupIntegrationTest) TestTargetClusterSelectedAutomatically() {
 	s.T().Logf("user signup '%s' created", userSignup.Name)
 
 	// Check the UserSignup is approved now
-	userSignup, err = s.hostAwait.WaitForUserSignup(userSignup.Name, wait.UntilUserSignupHasConditions(approvedAutomatically()...))
+	userSignup, err = s.hostAwait.WaitForUserSignup(userSignup.Name, wait.UntilUserSignupHasConditions(testsupport.ApprovedAutomatically()...))
 	require.NoError(s.T(), err)
 
 	// Confirm the MUR was created and target cluster was set
-	verifyResourcesProvisionedForSignup(s.T(), s.awaitility, *userSignup, "basic")
+	testsupport.VerifyResourcesProvisionedForSignup(s.T(), s.awaitility, *userSignup, "basic")
 }
 
 func (s *userSignupIntegrationTest) TestTransformUsername() {
 	// Create UserSignup with a username that we don't need to transform
-	userSignup, _ := s.createAndCheckUserSignup(true, "paul-no-need-to-transform", "paulnoneedtotransform@hotel.com", approvedByAdmin()...)
+	userSignup, _ := s.createAndCheckUserSignup(true, "paul-no-need-to-transform", "paulnoneedtotransform@hotel.com", testsupport.ApprovedByAdmin()...)
 	require.Equal(s.T(), "paul-no-need-to-transform", userSignup.Status.CompliantUsername)
 
 	// Create UserSignup with a username to transform
-	userSignup, _ = s.createAndCheckUserSignup(true, "paul@hotel.com", "paul@hotel.com", approvedByAdmin()...)
+	userSignup, _ = s.createAndCheckUserSignup(true, "paul@hotel.com", "paul@hotel.com", testsupport.ApprovedByAdmin()...)
 	require.Equal(s.T(), "paul", userSignup.Status.CompliantUsername)
 
 	// Create another UserSignup with the original username matching the transformed username of the existing signup
-	userSignup, _ = s.createAndCheckUserSignup(true, "paul", "paulathotel@hotel.com", approvedByAdmin()...)
+	userSignup, _ = s.createAndCheckUserSignup(true, "paul", "paulathotel@hotel.com", testsupport.ApprovedByAdmin()...)
 	require.Equal(s.T(), "paul-2", userSignup.Status.CompliantUsername)
 
 	// Create another UserSignup with the same original username but different user ID
-	userSignup, _ = s.createAndCheckUserSignup(true, "paul@hotel.com", "paul@hotel.com", approvedByAdmin()...)
+	userSignup, _ = s.createAndCheckUserSignup(true, "paul@hotel.com", "paul@hotel.com", testsupport.ApprovedByAdmin()...)
 	require.Equal(s.T(), "paul-3", userSignup.Status.CompliantUsername)
 }
 
@@ -111,14 +108,14 @@ func (s *userSignupIntegrationTest) createUserSignupAndAssertPendingApproval() *
 	// Create a new UserSignup with approved flag set to false
 	username := "testuser" + uuid.NewV4().String()
 	email := username + "@test.com"
-	userSignup := newUserSignup(s.T(), s.awaitility.Host(), username, email)
+	userSignup := testsupport.NewUserSignup(s.T(), s.awaitility.Host(), username, email)
 
 	err := s.awaitility.Client.Create(context.TODO(), userSignup, testsupport.CleanupOptions(s.ctx))
 	require.NoError(s.T(), err)
 	s.T().Logf("user signup '%s' created", userSignup.Name)
 
 	// Check the UserSignup is pending approval now
-	userSignup, err = s.hostAwait.WaitForUserSignup(userSignup.Name, wait.UntilUserSignupHasConditions(pendingApproval()...))
+	userSignup, err = s.hostAwait.WaitForUserSignup(userSignup.Name, wait.UntilUserSignupHasConditions(testsupport.PendingApproval()...))
 	require.NoError(s.T(), err)
 
 	// Confirm the CompliantUsername has NOT been set
@@ -132,12 +129,12 @@ func (s *userSignupIntegrationTest) createUserSignupAndAssertPendingApproval() *
 
 func (s *userSignupIntegrationTest) createUserSignupAndAssertManualApproval(specApproved bool) (*v1alpha1.UserSignup, *v1alpha1.MasterUserRecord) {
 	id := uuid.NewV4().String()
-	return s.createAndCheckUserSignup(specApproved, "testuser"+id, "testuser"+id+"@test.com", approvedByAdmin()...)
+	return s.createAndCheckUserSignup(specApproved, "testuser"+id, "testuser"+id+"@test.com", testsupport.ApprovedByAdmin()...)
 }
 
 func (s *userSignupIntegrationTest) createUserSignupAndAssertAutoApproval(specApproved bool) (*v1alpha1.UserSignup, *v1alpha1.MasterUserRecord) {
 	id := uuid.NewV4().String()
-	return s.createAndCheckUserSignup(specApproved, "testuser"+id, "testuser"+id+"@test.com", approvedAutomatically()...)
+	return s.createAndCheckUserSignup(specApproved, "testuser"+id, "testuser"+id+"@test.com", testsupport.ApprovedAutomatically()...)
 }
 
 func (s *userSignupIntegrationTest) checkUserSignupManualApproval() {
@@ -151,38 +148,15 @@ func (s *userSignupIntegrationTest) checkUserSignupManualApproval() {
 		require.NoError(s.T(), err)
 
 		// Check the UserSignup is approved now
-		userSignup, err = s.hostAwait.WaitForUserSignup(userSignup.Name, wait.UntilUserSignupHasConditions(approvedByAdmin()...))
+		userSignup, err = s.hostAwait.WaitForUserSignup(userSignup.Name, wait.UntilUserSignupHasConditions(testsupport.ApprovedByAdmin()...))
 		require.NoError(s.T(), err)
 
 		// Confirm the MUR was created
-		verifyResourcesProvisionedForSignup(s.T(), s.awaitility, *userSignup, "basic")
+		testsupport.VerifyResourcesProvisionedForSignup(s.T(), s.awaitility, *userSignup, "basic")
 	})
 
 	s.T().Run("usersignup created with approved set to true", func(t *testing.T) {
 		// Create a new UserSignup with approved flag set to true and assert approval
 		s.createUserSignupAndAssertManualApproval(true)
 	})
-}
-
-func newUserSignup(t *testing.T, host *wait.HostAwaitility, username string, email string) *v1alpha1.UserSignup {
-	memberCluster, ok, err := host.GetToolchainCluster(cluster.Member, wait.ReadyToolchainCluster)
-	require.NoError(t, err)
-	require.True(t, ok)
-
-	return &v1alpha1.UserSignup{
-		ObjectMeta: v1.ObjectMeta{
-			Name:      uuid.NewV4().String(),
-			Namespace: host.Ns,
-			Annotations: map[string]string{
-				v1alpha1.UserSignupUserEmailAnnotationKey: email,
-			},
-			Labels: map[string]string{
-				v1alpha1.UserSignupUserEmailHashLabelKey: md5.CalcMd5(email),
-			},
-		},
-		Spec: v1alpha1.UserSignupSpec{
-			Username:      username,
-			TargetCluster: memberCluster.Name,
-		},
-	}
 }
